C#: Present pointer fields as arrays.
Pointer struct members can be tagged with csharp_array to generate an
array field rather than a pointer.

Integer fields can be tagged with csharp_array_count(<array_field>) to
automatically keep the integer field in sync with the array count and
remove the burden of updating it manually.

Arrays generated in this manner are still serialized as regular
pointers.